runtime: extract telemetry ownership
