+++ crates/izwi-core/src/runtime/telemetry/mod.rs
@@ -1,0 +1,9 @@
+//! Runtime telemetry ownership.
+
+mod metrics;
+
+pub(crate) use metrics::{push_engine_metric, RuntimeTelemetryCollector};
+pub use metrics::{
+    EngineRuntimeTelemetrySnapshot, InferenceBrokerRuntimeTelemetrySnapshot,
+    PipelineRuntimeTelemetrySnapshot, RuntimeTelemetrySnapshot, VoiceRuntimeTelemetrySnapshot,
+};
